test(retrieval): cover worker abort listener cleanup
